Fix for multiple unknown number chats (#1979)
A quick fix when there are multiple unknown chats and auto-respond OOO hours turned on. Before this, the query would return the very first chat (which is usually wrong when there are multiple) with patientId = null. Now, we have added the comparison of patientPhoneNumber to ensure the right chat is returned.